+++ component/pom.xml
@@ -20,7 +20,7 @@
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <groupId>org.wso2.extension.siddhi.io.jms</groupId>
+        <groupId>io.siddhi.extension.io.jms</groupId>
         <artifactId>siddhi-io-jms-parent</artifactId>
         <version>2.0.2-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
@@ -93,7 +93,7 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.wso2.extension.siddhi.map.xml</groupId>
+            <groupId>io.siddhi.extension.map.xml</groupId>
             <artifactId>siddhi-map-xml</artifactId>
             <scope>test</scope>
         </dependency>
@@ -109,17 +109,17 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.wso2.extension.siddhi.map.json</groupId>
+            <groupId>io.siddhi.extension.map.json</groupId>
             <artifactId>siddhi-map-json</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.wso2.extension.siddhi.map.binary</groupId>
+            <groupId>io.siddhi.extension.map.binary</groupId>
             <artifactId>siddhi-map-binary</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.wso2.extension.siddhi.map.keyvalue</groupId>
+            <groupId>io.siddhi.extension.map.keyvalue</groupId>
             <artifactId>siddhi-map-keyvalue</artifactId>
             <scope>test</scope>
         </dependency>
@@ -179,7 +179,7 @@
                             org.wso2.carbon.kernel.utils.*
                         </Private-Package>
                         <Export-Package>
-                            org.wso2.extension.siddhi.io.jms.*
+                            io.siddhi.extension.io.jms.*
                         </Export-Package>
                         <Import-Package>
                             io.siddhi.core.*;version="${siddhi.version.range}",
