+++ docs/index.md
@@ -78,7 +78,7 @@ Latest API Docs is <a target="_blank" href="https://wso2-extensions.github.io/si
 
 ```
      <dependency>
-        <groupId>org.wso2.extension.siddhi.io.jms</groupId>
+        <groupId>io.siddhi.extension.io.jms</groupId>
         <artifactId>siddhi-io-jms</artifactId>
         <version>x.x.x</version>
      </dependency>
