+++ findbugs-exclude.xml
@@ -19,15 +19,15 @@
 
 <FindBugsFilter>
     <Match>
-        <class name="org.wso2.extension.siddhi.io.jms.source.JMSMessageProcessor"/>
+        <class name="io.siddhi.extension.io.jms.source.JMSMessageProcessor"/>
         <Bug pattern="BC_UNCONFIRMED_CAST"/>
     </Match>
     <Match>
-        <class name="org.wso2.extension.siddhi.io.jms.source.JMSMessageProcessor"/>
+        <class name="io.siddhi.extension.io.jms.source.JMSMessageProcessor"/>
         <Bug pattern="WA_AWAIT_NOT_IN_LOOP"/>
     </Match>
     <Match>
-        <class name="org.wso2.extension.siddhi.io.jms.source.JMSMessageProcessor"/>
+        <class name="io.siddhi.extension.io.jms.source.JMSMessageProcessor"/>
         <Bug pattern="EI_EXPOSE_REP2"/>
     </Match>
 </FindBugsFilter>
